feat(providers): reactivate AT&T and TalentoNet via shared residential proxy
AT&T's anti-bot blocked datacenter IPs and non-Mexican residential exits
alike; TalentoNet's server just wouldn't complete a TCP handshake through
datacenter proxies. Both work reliably through a residential proxy with a
Mexico exit — verified with repeated live lookups before re-enabling.

Extracted the proxy URL parsing (previously duplicated in att.ts and
freedompop.ts) into lib/proxy.ts, and pointed all three proxy-dependent
providers (AT&T, Freedompop, TalentoNet) at one RESIDENTIAL_PROXIES env
var instead of three separate ones. Dialo showed no rate limiting even
under load, so it's left making direct requests.
